+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.5.0] (Unreleased)
 - Added AnyPcapReader to read both pcap and pcapng files
+- Introduce AsyncPcapNgReader for asynchronous pcapng reading
+- Added AsyncAnyPcapReader for asynchronous reading of both pcap and pcapng files
 
 ## [0.4.0] (2026-01-06)
 - Fixed Bad Error Name